Renamed tests.common.MgmtSystem to mgmt_system
